#4082: Move SelectionToolPopover to subfolder
