+++ src/components/selectionToolPopover/SelectionToolPopover.module.scss
@@ -19,8 +19,6 @@ $background: rgb(252, 252, 252);
 $border: rgba(0, 0, 0, 0.3);
 
 .popover-wrapper {
-  width: fit-content;
-  height: fit-content;
   position: fixed;
   bottom: 30px;
   left: 30px;
